And we are back to where we started.
If our dependencies are dynamic (i.e. they depend on run-time values) we need to
keep track of them (i.e. store them somwhere). To determine if we should
recompute we determine if any of our dependencies is more recent than ourself.

When recomputing, we must update ourself to the most recent of our new
dependencies because that is the latest revision agains which we have computed
ourselves. Our new dependencies can be different from the old ones so we cannot
determine the new value of last_computed in the "should recompute" routine.